refactor: "enum" migrated to "enum class"

+++ PeltierChiller.Arduino/BME280.cpp
@@ -47,8 +47,8 @@ DynamicJsonDocument Models::TemperatureSensors::BME280::createPayload()
 {
 	DynamicJsonDocument document(_payloadSize);
 	JsonObject payload = document.to<JsonObject>();
-	payload["Type"] = getSensorType();
-	payload["Target"] = getSensorTarget();
+	payload["Type"] = static_cast<uint16_t>(getSensorType());
+	payload["Target"] = static_cast<uint16_t>(getSensorTarget());
 	payload["Temperature"] = getTemperature();
 	payload["Pressure"] = getPressure();
 	payload["Humidity"] = getHumidity();
